Organized the code.
