--- src/nl/tudelft/lifetiles/Main.fxml
@@ -1,70 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import java.lang.*?>
-<?import javafx.geometry.*?>
-<?import javafx.scene.*?>
-<?import javafx.scene.control.*?>
-<?import javafx.scene.layout.*?>
-<?import javafx.scene.text.*?>
-<?import javafx.scene.control.Label?>
-
-<VBox id="wrapper" prefHeight="600" prefWidth="800"
-	xmlns="http://javafx.com/javafx/8.0.40" xmlns:fx="http://javafx.com/fxml/1"
-	fx:controller="nl.tudelft.lifetiles.MenuController">
-	<MenuBar id="menubar" fx:id="menuBar" onKeyPressed="#handleKeyInput">
-		<menus>
-			<Menu text="File">
-				<items>
-					<MenuItem text="Open" />
-					<MenuItem text="Import" />
-					<SeparatorMenuItem />
-					<MenuItem text="Exit" />
-				</items>
-			</Menu>
-			<Menu text="View">
-				<items>
-					<MenuItem text="Foo" />
-					<MenuItem text="Bar" />
-				</items>
-			</Menu>
-			<Menu text="Configuration">
-				<items>
-					<MenuItem text="Foo" />
-					<MenuItem text="Bar" />
-				</items>
-			</Menu>
-			<Menu text="Help">
-				<items>
-					<MenuItem onAction="#handleAboutAction" text="About" />
-				</items>
-			</Menu>
-		</menus>
-	</MenuBar>
-	<ScrollPane fitToWidth="true" prefHeight="200.0" prefWidth="200.0">
-		<content>
-			<AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="200.0"
-				prefWidth="200.0">
-				<children>
-					<Label layoutX="7.0" layoutY="7.0" text="Graph Viewer" />
-				</children>
-			</AnchorPane>
-		</content>
-	</ScrollPane>
-	<SplitPane dividerPositions="0.2" prefHeight="160.0"
-		prefWidth="200.0">
-		<items>
-			<AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="160.0"
-				prefWidth="100.0">
-				<children>
-					<Label layoutX="7.0" layoutY="7.0" text="Data Viewer" />
-				</children>
-			</AnchorPane>
-			<AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="160.0"
-				prefWidth="100.0">
-				<children>
-					<Label layoutX="7.0" layoutY="7.0" text="Tree Viewer" />
-				</children>
-			</AnchorPane>
-		</items>
-	</SplitPane>
-</VBox>
