+++ src/nl/tudelft/lifetiles/controller/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * Contains the controllers.
+ * 
+ * @author Joren Hammudoglu
+ *
+ */
+package nl.tudelft.lifetiles.controller;
