+++ src/nl/tudelft/lifetiles/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * LifeTiles, the DNA visualizer. Main package.
+ * 
+ * @author Joren Hammudoglu
+ *
+ */
+package nl.tudelft.lifetiles;
