+++ src/nl/tudelft/lifetiles/view/DataStyle.css
@@ -1,0 +1,1 @@
+/* Stylesheet for the data view. */
