+++ src/nl/tudelft/lifetiles/view/GraphStyle.css
@@ -1,0 +1,1 @@
+/* Stylesheet for the graph view. */
