+++ src/nl/tudelft/lifetiles/view/MainStyle.css
@@ -1,0 +1,5 @@
+/* Stylesheet for the main view. */
+
+#wrapper {
+    -fx-background-color: #FFFFFF;
+}
