+++ src/nl/tudelft/lifetiles/view/MenuStyle.css
@@ -1,6 +1,4 @@
-#wrapper {
-    -fx-background-color: #FFFFFF;
-}
+/* Stylesheet for the menu view. */
 
 #menubar {
     -fx-background-color: #0B77B5;
