+++ src/nl/tudelft/lifetiles/view/MenuView.fxml
@@ -1,0 +1,40 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import java.net.*?>
+<?import javafx.scene.control.*?>
+
+<MenuBar xmlns:fx="http://javafx.com/fxml/1" id="menubar" fx:id="menuBar"
+	onKeyPressed="#handleKeyInput" fx:controller="nl.tudelft.lifetiles.controller.MenuController">
+	<menus>
+		<Menu text="File">
+			<items>
+				<MenuItem text="Open" />
+				<MenuItem text="Import" />
+				<SeparatorMenuItem />
+				<MenuItem text="Exit" />
+			</items>
+		</Menu>
+		<Menu text="View">
+			<items>
+				<MenuItem text="Foo" />
+				<MenuItem text="Bar" />
+			</items>
+		</Menu>
+		<Menu text="Configuration">
+			<items>
+				<MenuItem text="Foo" />
+				<MenuItem text="Bar" />
+			</items>
+		</Menu>
+		<Menu text="Help">
+			<items>
+				<MenuItem onAction="#handleAboutAction" text="About" />
+			</items>
+		</Menu>
+	</menus>
+
+	<stylesheets>
+		<URL value="@MenuStyle.css" />
+	</stylesheets>
+</MenuBar>
+
