+++ src/nl/tudelft/lifetiles/view/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * Contains the views.
+ * 
+ * @author Joren Hammudoglu
+ *
+ */
+package nl.tudelft.lifetiles.view;
